+++ src/frontend/apps/impress/src/features/docs/doc-management/stores/useProviderStore.tsx
@@ -1,0 +1,52 @@
+import { HocuspocusProvider } from '@hocuspocus/provider';
+import * as Y from 'yjs';
+import { create } from 'zustand';
+
+import { Base64 } from '@/features/docs/doc-management';
+
+export interface UseCollaborationStore {
+  createProvider: (
+    providerUrl: string,
+    storeId: string,
+    initialDoc?: Base64,
+  ) => HocuspocusProvider;
+  destroyProvider: () => void;
+  provider: HocuspocusProvider | undefined;
+}
+
+const defaultValues = {
+  provider: undefined,
+};
+
+export const useProviderStore = create<UseCollaborationStore>((set, get) => ({
+  ...defaultValues,
+  createProvider: (wsUrl, storeId, initialDoc) => {
+    const doc = new Y.Doc({
+      guid: storeId,
+    });
+
+    if (initialDoc) {
+      Y.applyUpdate(doc, Buffer.from(initialDoc, 'base64'));
+    }
+
+    const provider = new HocuspocusProvider({
+      url: wsUrl,
+      name: storeId,
+      document: doc,
+    });
+
+    set({
+      provider,
+    });
+
+    return provider;
+  },
+  destroyProvider: () => {
+    const provider = get().provider;
+    if (provider) {
+      provider.destroy();
+    }
+
+    set(defaultValues);
+  },
+}));